feat: update main view
