--- src/dashboard/WatchListColumn1Chart.jsx
@@ -1,102 +1,0 @@
-/* eslint-disable react/forbid-prop-types */
-/* eslint-disable react/jsx-props-no-spreading */
-import PropTypes from 'prop-types';
-
-import { alpha, useTheme } from '@mui/material';
-import React from 'react';
-
-const WatchListColumn1Chart = ({ data: dataProp, labels, ...rest }) => {
-    const theme = useTheme();
-
-    const data = (canvas) => {
-        const ctx = canvas.getContext('2d');
-        const primaryGradient = ctx.createLinearGradient(6, 6, 6, 150);
-
-        primaryGradient.addColorStop(0, alpha(theme.colors.primary.light, 0.8));
-        primaryGradient.addColorStop(0.8, theme.colors.alpha.white[10]);
-        primaryGradient.addColorStop(1, theme.colors.alpha.white[100]);
-
-        return {
-            datasets: [
-                {
-                    data: dataProp,
-                    borderWidth: 1,
-                    backgroundColor: primaryGradient,
-                    borderColor: theme.colors.primary.main,
-                    pointBorderWidth: 0,
-                    pointRadius: 0,
-                    pointHoverRadius: 0,
-                },
-            ],
-            labels,
-        };
-    };
-
-    const options = {
-        responsive: true,
-        maintainAspectRatio: false,
-        legend: {
-            display: false,
-        },
-        layout: {
-            padding: 0,
-        },
-        scales: {
-            xAxes: [
-                {
-                    gridLines: {
-                        display: false,
-                        drawBorder: false,
-                    },
-                    ticks: {
-                        display: false,
-                    },
-                },
-            ],
-            yAxes: [
-                {
-                    gridLines: {
-                        display: false,
-                    },
-                    ticks: {
-                        display: false,
-                    },
-                },
-            ],
-        },
-        tooltips: {
-            enabled: true,
-            mode: 'nearest',
-            intersect: false,
-            caretSize: 6,
-            displayColors: false,
-            yPadding: 8,
-            xPadding: 16,
-            borderWidth: 4,
-            borderColor: theme.palette.common.black,
-            backgroundColor: theme.palette.common.black,
-            titleFontColor: theme.palette.common.white,
-            bodyFontColor: theme.palette.common.white,
-            footerFontColor: theme.palette.common.white,
-            callbacks: {
-                title: () => {},
-                label: (tooltipItem) => {
-                    return `Price: $${tooltipItem.yLabel}`;
-                },
-            },
-        },
-    };
-
-    return (
-        <div {...rest}>
-            <Line data={data} options={options} />
-        </div>
-    );
-};
-
-WatchListColumn1Chart.propTypes = {
-    data: PropTypes.array.isRequired,
-    labels: PropTypes.array.isRequired,
-};
-
-export default WatchListColumn1Chart;
