+++ src/layouts/SidebarLayout/Header/Userbox/index.jsx
@@ -3,17 +3,17 @@ import React, { useRef, useState } from 'react';
 import { NavLink } from 'react-router-dom';
 
 import {
-  Avatar,
-  Box,
-  Button,
-  Divider,
-  Hidden,
-  lighten,
-  List,
-  ListItem,
-  ListItemText,
-  Popover,
-  Typography
+    Avatar,
+    Box,
+    Button,
+    Divider,
+    Hidden,
+    lighten,
+    List,
+    ListItem,
+    ListItemText,
+    Popover,
+    Typography,
 } from '@mui/material';
 
 import InboxTwoToneIcon from '@mui/icons-material/InboxTwoTone';
@@ -24,131 +24,117 @@ import LockOpenTwoToneIcon from '@mui/icons-material/LockOpenTwoTone';
 import AccountTreeTwoToneIcon from '@mui/icons-material/AccountTreeTwoTone';
 
 const UserBoxButton = styled(Button)(
-  ({ theme }) => `
+    ({ theme }) => `
         padding-left: ${theme.spacing(1)};
         padding-right: ${theme.spacing(1)};
-`
+`,
 );
 
 const MenuUserBox = styled(Box)(
-  ({ theme }) => `
+    ({ theme }) => `
         background: ${theme.colors.alpha.black[5]};
         padding: ${theme.spacing(2)};
-`
+`,
 );
 
 const UserBoxText = styled(Box)(
-  ({ theme }) => `
+    ({ theme }) => `
         text-align: left;
         padding-left: ${theme.spacing(1)};
-`
+`,
 );
 
 const UserBoxLabel = styled(Typography)(
-  ({ theme }) => `
+    ({ theme }) => `
         font-weight: ${theme.typography.fontWeightBold};
         color: ${theme.palette.secondary.main};
         display: block;
-`
+`,
 );
 
 const UserBoxDescription = styled(Typography)(
-  ({ theme }) => `
+    ({ theme }) => `
         color: ${lighten(theme.palette.secondary.main, 0.5)}
-`
+`,
 );
 
 const HeaderUserbox = () => {
+    const user = {
+        name: 'User',
+        avatar: '',
+        jobtitle: 'Profession',
+    };
 
-  const user =
-  {
-    name: 'Catherine Pike',
-    avatar: '/static/images/avatars/1.jpg',
-    jobtitle: 'Project Manager'
-  };
+    const ref = useRef(null);
+    const [isOpen, setOpen] = useState(false);
 
-  const ref = useRef(null);
-  const [isOpen, setOpen] = useState(false);
+    const handleOpen = () => {
+        setOpen(true);
+    };
 
-  const handleOpen = () => {
-    setOpen(true);
-  };
+    const handleClose = () => {
+        setOpen(false);
+    };
 
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <React.Fragment>
-      <UserBoxButton color="secondary" ref={ref} onClick={handleOpen}>
-        <Avatar variant="rounded" alt={user.name} src={user.avatar} />
-        <Hidden mdDown>
-          <UserBoxText>
-            <UserBoxLabel variant="body1">{user.name}</UserBoxLabel>
-            <UserBoxDescription variant="body2">
-              {user.jobtitle}
-            </UserBoxDescription>
-          </UserBoxText>
-        </Hidden>
-        <Hidden smDown>
-          <ExpandMoreTwoToneIcon sx={{ ml: 1 }} />
-        </Hidden>
-      </UserBoxButton>
-      <Popover
-        anchorEl={ref.current}
-        onClose={handleClose}
-        open={isOpen}
-        anchorOrigin={{
-          vertical: 'top',
-          horizontal: 'right'
-        }}
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'right'
-        }}
-      >
-        <MenuUserBox sx={{ minWidth: 210 }} display="flex">
-          <Avatar variant="rounded" alt={user.name} src={user.avatar} />
-          <UserBoxText>
-            <UserBoxLabel variant="body1">{user.name}</UserBoxLabel>
-            <UserBoxDescription variant="body2">
-              {user.jobtitle}
-            </UserBoxDescription>
-          </UserBoxText>
-        </MenuUserBox>
-        <Divider sx={{ mb: 0 }} />
-        <List sx={{ p: 1 }} component="nav">
-          <ListItem button to="/management/profile/details" component={NavLink}>
-            <AccountBoxTwoToneIcon fontSize="small" />
-            <ListItemText primary="My Profile" />
-          </ListItem>
-          <ListItem
-            button
-            to="/dashboards/messenger"
-            component={NavLink}
-          >
-            <InboxTwoToneIcon fontSize="small" />
-            <ListItemText primary="Messenger" />
-          </ListItem>
-          <ListItem
-            button
-            to="/management/profile/settings"
-            component={NavLink}
-          >
-            <AccountTreeTwoToneIcon fontSize="small" />
-            <ListItemText primary="Account Settings" />
-          </ListItem>
-        </List>
-        <Divider />
-        <Box sx={{ m: 1 }}>
-          <Button color="primary" fullWidth>
-            <LockOpenTwoToneIcon sx={{ mr: 1 }} />
-            Sign out
-          </Button>
-        </Box>
-      </Popover>
-    </React.Fragment>
-  );
-}
+    return (
+        <React.Fragment>
+            <UserBoxButton color="secondary" ref={ref} onClick={handleOpen}>
+                <Avatar variant="rounded" alt={user.name} />
+                <Hidden mdDown>
+                    <UserBoxText>
+                        <UserBoxLabel variant="body1">{user.name}</UserBoxLabel>
+                        <UserBoxDescription variant="body2">{user.jobtitle}</UserBoxDescription>
+                    </UserBoxText>
+                </Hidden>
+                <Hidden smDown>
+                    <ExpandMoreTwoToneIcon sx={{ ml: 1 }} />
+                </Hidden>
+            </UserBoxButton>
+            <Popover
+                anchorEl={ref.current}
+                onClose={handleClose}
+                open={isOpen}
+                anchorOrigin={{
+                    vertical: 'top',
+                    horizontal: 'right',
+                }}
+                transformOrigin={{
+                    vertical: 'top',
+                    horizontal: 'right',
+                }}
+            >
+                <MenuUserBox sx={{ minWidth: 210 }} display="flex">
+                    <Avatar variant="rounded" alt={user.name} src={user.avatar} />
+                    <UserBoxText>
+                        <UserBoxLabel variant="body1">{user.name}</UserBoxLabel>
+                        <UserBoxDescription variant="body2">{user.jobtitle}</UserBoxDescription>
+                    </UserBoxText>
+                </MenuUserBox>
+                <Divider sx={{ mb: 0 }} />
+                <List sx={{ p: 1 }} component="nav">
+                    <ListItem button to="/management/profile/details" component={NavLink}>
+                        <AccountBoxTwoToneIcon fontSize="small" />
+                        <ListItemText primary="My Profile" />
+                    </ListItem>
+                    <ListItem button to="/dashboards/messenger" component={NavLink}>
+                        <InboxTwoToneIcon fontSize="small" />
+                        <ListItemText primary="Messenger" />
+                    </ListItem>
+                    <ListItem button to="/management/profile/settings" component={NavLink}>
+                        <AccountTreeTwoToneIcon fontSize="small" />
+                        <ListItemText primary="Account Settings" />
+                    </ListItem>
+                </List>
+                <Divider />
+                <Box sx={{ m: 1 }}>
+                    <Button color="primary" fullWidth>
+                        <LockOpenTwoToneIcon sx={{ mr: 1 }} />
+                        Sign out
+                    </Button>
+                </Box>
+            </Popover>
+        </React.Fragment>
+    );
+};
 
 export default HeaderUserbox;
